feat: Add ParticleTypesRegistry
There is no need to put registry code onto the platform mods, registering is a minecraft dependent process, you can register in common and use the platform method as an interface. The way this particle registry is implemented will avoid duplicate code in the platforms as they won't need to be registered everywhere.